Update README.md
A Python binary search finds the position of an item in a sorted array. It divides the list in half. If a specified value is higher than the middle number, the search focuses on the right of the list. Otherwise, the search looks for the number on the left of the list. 

This python project idea will help create an algorithm that searches for an element in a list. 

There are two ways which we can perform this search;
	Iterative method (makes use of the while loop)
Recursive method (uses the divide and conquer method)